sql_lite_dbのテストを分割
SQLコマンド発行とそれ以外へファイル分離。
合わせてテストも分離。
